dictaphone/requirements.txt
pyaudio==0.2.11
gTTS==2.2.3
SpeechRecognition==3.8.1
python-socketio==5.4.0
eventlet==0.30.2